pyproject.toml
[tool.poetry]
name = "retailchain-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
coverage = "^7.6.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
